Refactor prober dependency
